+++ generators/app/templates/src/containers/Profile/components/Form/form.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, cleanup } from 'react-testing-library';
-import { FormUi } from "./form.component.ui";
+import { FormUi } from "./form.component";
 
 
 describe("Profile Container", () => {
