+++ generators/app/templates/src/containers/Profile/components/Form/index.js
@@ -1,3 +1,3 @@
-import Form from './form.component';
+import Form from './form.container';
 
 export default Form;
